\name{build_proto}
\alias{build_proto}
\title{builds proto widgets}
\description{The function build_proto is called to build the proto widgets}
\author{cedric}
\note{Unlike S3 or S4 class, the structure of proto objects is lost when the package is built
so we embedded them in a function which will be called to create them.}
\references{\url{http://www.docstoc.com/docs/2558547/gWidgets-Example-Proto-example-2-This-example-shows-how-to-use-} especially ProtoExample}
