\name{chnames}
\alias{chnames}
\title{This function replaces the variable names in a data.frame}
\usage{
  chnames(objet, old_variable_name, new_variable_name)
}
\arguments{
  \item{objet}{a data frame}

  \item{old_variable_name}{}

  \item{new_variable_name}{}
}
\value{
  objet
}
\description{
  This function replaces the variable names in a data.frame
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

