\name{confirmDialog}
\alias{confirmDialog}
\title{Confirmation dialog gtkwindow when selecting a different dataframe...}
\usage{confirmDialog(message, handlerok, handlercancel)}
\description{Confirmation dialog gtkwindow when selecting a different dataframe
This is directly taken from gWidget vignette and slightly modified}
\author{John Verzani}
\arguments{
\item{handlerok}{the handler triggered when ok is clicked}
\item{handlercancel}{the handler triggered when cancel is clicked}}