% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionDB.R
\name{connect,ConnectionDB-method}
\alias{connect,ConnectionDB-method}
\alias{connect.ConnectionDB}
\title{connect method for ConnectionDB class}
\usage{
\S4method{connect}{ConnectionDB}(object, base = NULL)
}
\arguments{
\item{object}{An object of class ConnectionDB}

\item{base}{a string with values for dbname, host, port, user, password, in this order.}
}
\value{
a connection with slot filled
}
\description{
connect method for ConnectionDB class
}
\examples{
\dontrun{
object <- new("ConnectionDB")
object@dbname <- c("bd_contmig_nat","test")
object@host <- 		"localhost"
object@port <-		"5432"
object@user <-		"myuser"
object@password <-		"mypassword"
object@silent=FALSE
object <- connect(object)
pool::dbGetInfo(object@connection)
pool::poolClose(object@connection)
}
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
