% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionODBC.R
\name{connect,ConnectionODBC-method}
\alias{connect,ConnectionODBC-method}
\alias{connect.ConnectionODBC}
\title{connect method for ConnectionODBC class}
\usage{
\S4method{connect}{ConnectionODBC}(object)
}
\arguments{
\item{object}{an object of class ConnectionODBC}
}
\value{
a connection with slot filled
}
\description{
connect method for ConnectionODBC class
}
\examples{
##this wont be run as the user need to manually configure the connection before using it
\dontrun{
object=new("ConnectionODBC")
object@baseODBC=c("myODBCconnection","myusername","mypassword")
object@silent=FALSE
object<-connect(object)
odbcClose(object@connection)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
