% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequeteODBCwhere.R
\name{connect,RequeteODBCwhere-method}
\alias{connect,RequeteODBCwhere-method}
\alias{connect.RequeteODBCwhere}
\title{connect method loads a request to the database and returns either an error or a data.frame}
\usage{
\S4method{connect}{RequeteODBCwhere}(object)
}
\arguments{
\item{object}{an object of class RequeteODBCwhere}
}
\value{
An object of class RequeteODBCwhere
}
\description{
connect method loads a request to the database and returns either an error or a data.frame
}
\note{
method modified from v0.2.1240 to use the connect method of the mother class
}
\examples{
\dontrun{
object<-new("RequeteODBCwhere")
object@baseODBC<-baseODBC
object@sql<- "select * from t_lot_lot"
object@where<-"WHERE lot_tax_code='2038'"
object@and<-c("AND lot_std_code='CIV'","AND lot_ope_identifiant<1000")
object@order_by<-"ORDER BY lot_identifiant"
object<-connect(object)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
