% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequeteODBCwheredate.R
\name{connect,RequeteODBCwheredate-method}
\alias{connect,RequeteODBCwheredate-method}
\alias{connect.RequeteODBCwheredate}
\title{connect method loads a request to the database and returns either an error or a data.frame}
\usage{
\S4method{connect}{RequeteODBCwheredate}(object)
}
\arguments{
\item{object}{an object of class RequeteODBCwheredate}
}
\value{
An object of class RequeteODBCwheredate
}
\description{
connect method loads a request to the database and returns either an error or a data.frame
}
\note{
method modified from v0.2.1240 to use the connect method of the mother class which in turn will use the method of the mother class
}
\examples{
\dontrun{
object<-new("RequeteODBCwheredate")
object@baseODBC<-get("baseODBC",envir=envir_stacomi)
object@select<- "select * from t_operation_ope"
object@datedebut=strptime("1996-01-01 00:00:00",format="\%Y-\%m-\%d \%H:\%M:\%S")
object@datefin=strptime("2000-01-01 00:00:00",format="\%Y-\%m-\%d \%H:\%M:\%S")
object@colonnedebut="ope_date_debut"
object@colonnefin="ope_date_fin"
object@and<-c("AND ope_dic_identifiant=1","AND ope_dic_identifiant=2")
object@order_by<-"ORDER BY ope_identifiant"
object@silent=FALSE
object<-connect(object) 
}
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
