\name{connect}
\alias{connect}
\title{connect method for ConnexionODBC class}


\description{
  connect method loads a request to the database and
  returns either an error or a data.frame 
}

\examples{\dontrun objet<-new("RequeteODBCwhere")
connext(objet)
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}


