% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionODBC.R
\docType{class}
\name{ConnectionODBC-class}
\alias{ConnectionODBC-class}
\title{ConnectionODBC class}
\value{
connectionODBC an S4 object of class connectionODBC
}
\description{
ConnectionODBC class
}
\section{Slots}{

\describe{
\item{\code{baseODBC}}{"vector" (of length 3, character)}

\item{\code{silent}}{"logical"}

\item{\code{etat}}{"ANY" # can be -1 or string}

\item{\code{connection}}{"ANY" # could be both string or S3}
}}

\note{
Mother class for connection, opens the connection but does not shut it
}
\examples{
##this wont be run as the user need to manually configure the connection before using it
\dontrun{
object=new("ConnectionODBC")
object@baseODBC=c("myODBCconnection","myusername","mypassword")
object@silent=FALSE
object<-connect(object)
odbcClose(object@connection)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
