\name{ex}
\alias{ex}
\title{ex fonction to write to excel, not used within the program but can still be used}
\usage{
  ex(d = NULL)
}
\arguments{
  \item{d}{}
}
\description{
  ex fonction to write to excel, not used within the
  program but can still be used
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

