% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{funout}
\alias{funout}
\title{Function loaded in this package to avoid errors, if the package is called without stacomiR}
\usage{
funout(text, arret = FALSE, wash = FALSE)
}
\arguments{
\item{text}{The text to display}

\item{arret}{Boolean should the program stop}

\item{wash}{FALSE only used when called from within stacomiR, and there is a widget interface,
kept there for consistency}
}
\description{
Function loaded in this package to avoid errors, if the package is called without stacomiR
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
