% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequeteDB.R
\name{getquery,RequeteDB-method}
\alias{getquery,RequeteDB-method}
\title{getquery retreives the result of the query from the object}
\usage{
\S4method{getquery}{RequeteDB}(object)
}
\arguments{
\item{object}{an object of class RequeteDB}
}
\value{
A data frame
}
\description{
getquery retreives the result of the query from the object
}
