% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequeteDB.R
\name{getquery}
\alias{getquery}
\title{generic query function for}
\usage{
getquery(object, ...)
}
\arguments{
\item{object}{an object}

\item{...}{additional parameters passed to query}
}
\description{
generic query function for
}
