\name{haes}
\alias{haes}
\title{this handler first analyses the action (which is the name of the BoxinLayout)...}
\usage{haes(h, ...)}
\description{this handler first analyses the action (which is the name of the BoxinLayout)
then calls the load_aes function}
\author{cedric}
\arguments{\item{h}{a handler}}