\name{hChangedata}
\alias{hChangedata}
\title{This function will check if data has the same column than previously}
\usage{hChangedata(h, ...)}
\description{This function will check if data has the same column than previously, if not
It will rebuild the graphical interface with the new data}
\author{cedric}
\arguments{\item{h}{a handler for gdroplist data}}