\name{hretablir}
\alias{hretablir}
\title{This function restarts with the default geom, when button Undo is pushed...}
\usage{hretablir(h, ...)}
\description{This function restarts with the default geom, when button Undo is pushed}
\author{cedric}
\arguments{
\item{h}{handler}
\item{...}{other arguments to be passed to the function}
}