\name{hsw}
\alias{hsw}
\title{handler allowing the switch between stat and geom when button geom<>stat is triggered...}
\usage{hsw(h, ...)}
\description{handler allowing the switch between stat and geom when button geom<>stat is triggered}
\arguments{
\item{h}{handler}
\item{...}{other arguments to be passed to the function}
}
\author{cedric}
