\name{hUpdatedata}
\alias{hUpdatedata}
\title{This handler updates the list of data avalaible for drawing the graph...}
\usage{hUpdatedata(h, ...)}
\description{This handler updates the list of data avalaible for drawing the graph
If you want for instance to do some calculations in the dataframe and then reload it
It uses the add and delete methods applicable to gframe to change the content of the combobox by re-building it}
\author{cedric}
\arguments{
\item{h}{handler for gdroplist data}
\item{...}{other arguments to be passed to the function}
}