\name{ind}
\alias{ind}
\title{fonction pour renvoyer les index dans b  des valeurs du vecteur a
b peut apparaitre plusieurs fois dans a}
\usage{
  ind(a, b)
}
\arguments{
  \item{a}{}

  \item{b}{}
}
\value{
  index of b in a
}
\description{
  fonction pour renvoyer les index dans b des valeurs du
  vecteur a b peut apparaitre plusieurs fois dans a
}
\note{
  attention le vecteur de resultat est dans le desordre
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

