\name{indrepeated}
\alias{indrepeated}
\title{index of repeated values}
\usage{
  indrepeated(a)
}
\arguments{
  \item{a}{a vector}
}
\value{
  the index of repeated values within a vector (this function returns the index of all values not returned by induk)
}
\description{
  returns the index of repeated values within a vector
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

