\name{induk}
\alias{induk}
\title{fonction qui renvoit l'index des valeurs apparaissant une seule fois}
\usage{
  induk(a)
}
\arguments{
  \item{a}{}
}
\value{
  the index unique values within a vector
}
\description{
  fonction qui renvoit l'index des valeurs apparaissant une
  seule fois
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

