\name{is.even}
\alias{is.even}
\title{is.even function modified from package sma (which did not verified that the entry was indeed an integer)}
\usage{
  is.even(x)
}
\arguments{
  \item{x}{integer}
}
\value{
  a logical
}
\description{
  is.even function modified from package sma (which did not
  verified that the entry was indeed an integer)
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

