\name{is.odd}
\alias{is.odd}
\title{id.odd function modified from package sma (which did not verified that the entry was indeed an integer)}
\usage{
  is.odd(x)
}
\arguments{
  \item{x}{integer}
}
\value{
  a logical
}
\description{
  id.odd function modified from package sma (which did not
  verified that the entry was indeed an integer)
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

