\name{killfactor}
\alias{killfactor}
\title{very usefull function used to "kill" these bloody factors that appears, noticeably after loading with odbc}
\usage{
  killfactor(df)
}
\arguments{
  \item{df}{a data.frame}
}
\value{
  df
}
\description{
  very usefull function used to "kill" these bloody factors
  that appears, noticeably after loading with odbc
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

