\name{load_aes}
\alias{load_aes}
\title{this functions loads the graphical interface with default elements from layer...}
\usage{load_aes(layer_choice)}
\description{this functions loads the graphical interface with default elements from layer}
\author{cedric}
\arguments{\item{layer_choice}{either "geom" or "stat"}}
