\name{ls.class}
\alias{ls.class}
\title{Function used to list all elements belonging to a class within an environment...}
\usage{ls.class(name=.GlobalEnv, all.names=FALSE, pattern, class=data.frame)}
\description{Function used to list all elements belonging to a class within an environment}
\value{vector containing elements listed as belonging to a class}
\author{cedric}
\arguments{\item{name}{the name of the environment}
\item{all.names}{see ls() for use}
\item{pattern}{an optional regular expression see ls() for use}
\item{class}{An S3 class}}
