% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequeteDB.R
\name{query,RequeteDB-method}
\alias{query,RequeteDB-method}
\title{query method loads a quert to the data and returns either an error or a data.frame}
\usage{
\S4method{query}{RequeteDB}(object, ...)
}
\arguments{
\item{object}{an object of class RequeteDB}

\item{...}{further arguments passed to the query method, base will be passed to ConnectionDB to set the connection parameters,
it should be a vector with dbname host port user and password (in this order)}
}
\value{
An object of class RequeteDB
}
\description{
query method loads a quert to the data and returns either an error or a data.frame
}
\note{
assign("showmerequest",1,envir=envir_stacomi) allows to print all queries passing on the class RequeteDB
}
\examples{
showClass("RequeteDB")
\dontrun{
# this connection require user and password and a working data
 object=new("RequeteDB",dbname ="bd_contmig_nat",
			host ="localhost", port="5432", user ="postgres", password="secret")
)
object@open=TRUE
 ## this will leave the connection open, 
 ## by default it closes after the query is sent
 ## the following will work only if you have configured and ODBC link
object@sql= "select * from t_lot_lot limit 100"
object<-query(object)
  envir_stacomi=new.env()
  ## While testing if you want to see the output of sometimes
  ## complex queries generated by the program
  assign("showmerequest",1,envir_stacomi) 
  ## You can assign any values (here 1)
  ## just tests the existence of "showmerequest" in envir_stacomi
  object@sql= "select * from mytable limit 100"
  object<-connect(object)
 ## the connection is already closed, the query is printed
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
