% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequeteDBwhere.R
\name{query,RequeteDBwhere-method}
\alias{query,RequeteDBwhere-method}
\alias{query.RequeteODBCwhere}
\title{query method loads a request to the database and returns either an error or a data.frame}
\usage{
\S4method{query}{RequeteDBwhere}(object, ...)
}
\arguments{
\item{object}{an object of class RequeteDBwhere}

\item{...}{further arguments passed to the query method, base will be passed to ConnectionDB to set the connection parameters,}
}
\value{
An object of class RequeteODBCwhere
}
\description{
query method loads a request to the database and returns either an error or a data.frame
}
\examples{
\dontrun{
object<-new("RequeteODBCwhere")
base=c("bd_contmig_nat","localhost","5432","user", "password")
object@sql<- "select * from t_lot_lot"
object@where<-"WHERE lot_tax_code='2038'"
object@and<-c("AND lot_std_code='CIV'","AND lot_ope_identifiant<1000")
object@order_by<-"ORDER BY lot_identifiant"
object <- connect(object, base)
}
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
