% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequeteDBwheredate.R
\name{query,RequeteDBwheredate-method}
\alias{query,RequeteDBwheredate-method}
\alias{query.RequeteDBwheredate}
\title{query method loads a request to the database and returns either an error or a data.frame}
\usage{
\S4method{query}{RequeteDBwheredate}(object, ...)
}
\arguments{
\item{object}{an object of class RequeteDBwheredate}

\item{...}{further arguments passed to the query method, base will be passed to ConnectionDB to set the connection parameters,}
}
\value{
An object of class RequeteDBwheredate
}
\description{
query method loads a request to the database and returns either an error or a data.frame
}
\examples{
\dontrun{
object<-new("RequeteDBwheredate")
base=c("bd_contmig_nat","localhost","5432","user", "password")
object@select<- "select * from t_operation_ope"
object@datedebut=strptime("1996-01-01 00:00:00",format="\%Y-\%m-\%d \%H:\%M:\%S")
object@datefin=strptime("2000-01-01 00:00:00",format="\%Y-\%m-\%d \%H:\%M:\%S")
object@colonnedebut="ope_date_debut"
object@colonnefin="ope_date_fin"
object@and<-c("AND ope_dic_identifiant=1","AND ope_dic_identifiant=2")
object@order_by<-"ORDER BY ope_identifiant"
object@silent=FALSE
object<-connect(object, base)
}
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
