% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequeteDB.R
\name{query}
\alias{query}
\alias{query.RequeteDB}
\title{generic query function for}
\usage{
query(object, ...)
}
\arguments{
\item{object}{an object}

\item{...}{additional parameters passed to query}
}
\description{
generic query function for
}
