% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequeteDB.R
\docType{class}
\name{RequeteDB-class}
\alias{RequeteDB-class}
\title{RequeteDB class}
\description{
RequeteDB class
}
\section{Slots}{

\describe{
\item{\code{sql}}{"character"}

\item{\code{query}}{"data.frame"}

\item{\code{open}}{logical is the connection left open after the request ?}
}}

\note{
Inherits from ConnectionDB
}
\examples{
 object=new("RequeteDB")
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
