% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequeteDBwhere.R
\docType{class}
\name{RequeteDBwhere-class}
\alias{RequeteDBwhere-class}
\title{RequeteDBwhere class}
\description{
RequeteDBwhere class
}
\section{Slots}{

\describe{
\item{\code{select}}{"character"}

\item{\code{where}}{"character"}

\item{\code{and}}{"vector"}

\item{\code{order_by}}{"character"}
}}

\note{
Inherits from RequeteDB
the syntax is where="WHERE ..."
and =vector("AND...","AND...")
order_by="ORDER BY.."
}
\examples{
 object=new("RequeteDBwhere")
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
