% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequeteDBwheredate.R
\docType{class}
\name{RequeteDBwheredate-class}
\alias{RequeteDBwheredate-class}
\title{RequeteDBwhere class}
\description{
RequeteDBwhere class
}
\section{Slots}{

\describe{
\item{\code{datedebut}}{"POSIXlt"}

\item{\code{datefin}}{"POSIXlt"}

\item{\code{colonnedebut}}{"character" # name of the column containing datedebut}

\item{\code{colonnefin}}{"character"  # name of the column containing datefin}
}}

\note{
Inherits from RequeteDBwhere and uses its connect method with a new SetAs
}
\examples{
object=new("RequeteDBwhere")
}
