% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{stacomirtools}
\alias{stacomirtools}
\alias{stacomirtools-package}
\title{Stacomirtools options.}
\description{
S4 class wrappers for the 'ODBC' and Pool DBI connection, also provides some utilities to paste small datasets to clipboard, rename columns. It is used by the package 'stacomiR' for connections to the database. Development versions of 'stacomiR' are available in R-forge.
}
\section{Package options}{


stacomirtools uses the following [options()] to configure behaviour:

\itemize{
  \item `stacomiR.dbname`: databasename

  \item `stacomiR.host`: the name of the host, often ["localhost"]

  \item `stacomiR.port`: the name of the port, most often ["5432"]
   
  \item `stacomiR.user`: a string with the user name e.g. ["postgres"]#'     

  \item `stacomiR.password`: a string with the user password

  \item `stacomiR.ODBClink`: a string with name of the ODBC link, default NULL

  \item `stacomiR.path`: a string with the path to where some output are written  

   \item `stacomiR.printqueries`: a boolean, default FALSE, with the side effect of printing queries to the console
}
}

\author{
\strong{Maintainer}: Cedric Briand \email{cedric.briand00@gmail.com}

Authors:
\itemize{
  \item Marion Legrand \email{tableau-salt-loire@logrami.fr}
}

Other contributors:
\itemize{
  \item Beaulaton Laurent \email{laurent.beaulaton@ofb.gouv.fr} [contributor]
}

}
\keyword{internal}
