% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionDB.R
\name{validity_DB}
\alias{validity_DB}
\title{validity function for ConnectionDB class}
\usage{
validity_DB(object)
}
\arguments{
\item{object}{An object of class ConnectionDB}
}
\description{
validity function for ConnectionDB class
}
