% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionODBC.R
\name{validity_ODBC}
\alias{validity_ODBC}
\title{Validity method for ODBC class}
\usage{
validity_ODBC(object)
}
\arguments{
\item{object}{an object of class ConnectionODBC}
}
\description{
Validity method for ODBC class
}
