\name{chnames}
\alias{chnames}
\title{This function replaces the variable names in a data.frame}
\usage{
  chnames(object, old_variable_name, new_variable_name)
}
\arguments{
  \item{object}{a data frame}

  \item{old_variable_name}{a character vector with old variables names}

  \item{new_variable_name}{a character vector with new variables names}
}
\value{
  object
}
\description{
  This function replaces the variable names in a data.frame
}
\author{
  Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

