% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionODBC.R
\name{connect}
\alias{connect}
\alias{connect.generic}
\title{generic connect function for baseODBC}
\usage{
connect(object, ...)
}
\arguments{
\item{object}{an object}

\item{...}{additional arguments passed on to the connect method}
}
\description{
generic connect function for baseODBC
}
