% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ex}
\alias{ex}
\title{ex fonction to write to the clipboard}
\usage{
ex(d = NULL)
}
\arguments{
\item{d}{a dataframe}
}
\description{
ex fonction to write to the clipboard
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
