\name{indrepeated}
\alias{indrepeated}
\title{fonction qui retourne l'index des valeurs repetees d'un vecteur}
\usage{
  indrepeated(a)
}
\arguments{
  \item{a}{}
}
\value{
  the index of repeated values within a vector
}
\description{
  fonction qui retourne l'index des valeurs repetees d'un
  vecteur
}
\author{
  Cedric Briand \email{cedric.briand00@gmail.com}
}

