\name{induk}
\alias{induk}
\title{unique values of a vector}
\usage{
  induk(a)
}
\arguments{
  \item{a}{a vector}
}
\value{
  the index unique values within a vector
}
\description{
returns the index of values appearing only once in a vector : match(unique(a),a), replicated values are not returned on their second occurence
}
\author{
  Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

