\name{is.even}
\alias{is.even}
\title{is.even function modified from package sma}
\usage{
  is.even(x)
}
\arguments{
  \item{x}{integer}
}
\value{
  a logical
}
\description{
  is.even function modified from package sma (which did not
  verify that the entry was indeed an integer)
}
\author{
  Adapted from Henrik Bengtsson       

}

