\name{is.odd}
\alias{is.odd}
\title{id.odd function modified from package sma}
\usage{
  is.odd(x)
}
\arguments{
  \item{x}{integer}
}
\value{
  a logical
}
\description{
  id.odd function modified from package sma (which did not
  verify that the entry was indeed an integer)
}
\author{
  Adapted from Henrik Bengtsson       
}
\examples{
is.odd(1)
is.odd(2)
}
