% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{killfactor}
\alias{killfactor}
\title{very usefull function used to "kill" the factors, noticeably after loading with 'ODBC'}
\usage{
killfactor(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
df
}
\description{
very usefull function used to "kill" the factors, noticeably after loading with 'ODBC'
}
\author{
Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}
