\name{killfactor}
\alias{killfactor}
\title{very usefull function remove factor that appear, noticeably after loading with ODBC}
\usage{
  killfactor(df)
}
\arguments{
  \item{df}{a data.frame}
}
\value{
  df
}
\description{
  function used to remove factors that appear, noticeably after loading with ODBC
}
\author{
  Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}
}

