% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequeteODBC.R
\docType{class}
\name{RequeteODBC-class}
\alias{RequeteODBC-class}
\title{RequeteODBC class}
\description{
RequeteODBC class
}
\section{Slots}{

\describe{
\item{\code{sql}}{"character"}

\item{\code{query}}{"data.frame"}

\item{\code{open}}{logical is the connection left open after the request ?}
}}

\note{
Inherits from ConnectionODBC
}
\examples{
 object=new("RequeteODBC")
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
