% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequeteODBCwhere.R
\docType{class}
\name{RequeteODBCwhere-class}
\alias{RequeteODBCwhere-class}
\title{RequeteODBCwhere class}
\description{
RequeteODBCwhere class
}
\section{Slots}{

\describe{
\item{\code{select}}{"character"}

\item{\code{where}}{"character"}

\item{\code{and}}{"vector"}

\item{\code{order_by}}{"character"}
}}

\note{
Inherits from RequeteODBC
the syntax is where="WHERE ..."
and =vector("AND...","AND...")
order_by="ORDER BY.."
}
\examples{
 object=new("RequeteODBCwhere")
}
\author{
Cedric Briand \email{cedric.briand@eptb-vilaine.fr}
}
