\name{stacomirtools-package}
\title{RODBC connector class and some utilities}
\description{
This package contains S4 wrappers for 'ODBC' connection and some utilities
}
\details{\tabular{ll}{
Package: \tab stacomirtools\cr
Type: \tab Package\cr
Version: \tab 0.5.2\cr
Date: \tab 2017-06-24\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\docType{package}
\author{Cedric Briand \email{cedric.briand"at"eptb-vilaine.fr}}
\keyword{package}